refactor: 토큰 uri 변경 & 클래스 분리
