+++ security/src/main/resources/auth/application.yml
@@ -22,4 +22,4 @@ spring:
 
 token:
   request:
-    uri: '/token'
+    uri: '/tokens'
